完成

+++ src/components/TrashDetail.vue
@@ -54,8 +54,13 @@ export default {
     this.getTrashNotes()
       .then(() => {
         this.setCurTrashNote({ curTrashNoteId: this.$route.query.noteId })
+        this.$router.replace({
+          path: '/trash',
+          query: { noteId: this.curTrashNote.id }
+        })
       })
   },
+
 
   computed: {
     ...mapGetters([
@@ -83,12 +88,34 @@ export default {
       ]),
 
     onDelete() {
-      console.log({ noteId: this.curTrashNote.id })
-      this.deleteTrashNote({ noteId: this.curTrashNote.id })
+       this.$confirm('删除后将无法恢复', '确定删除？', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+           return this.deleteTrashNote({ noteId: this.curTrashNote.id })
+        }).then(()=>{
+            this.setCurTrashNote()
+            this.$router.replace({
+            path:'/trash',
+            query:{
+              noteId:this.curTrashNote.id,
+            }
+          })
+        })
     },
 
     onRevert() {
       this.revertTrashNote({ noteId: this.curTrashNote.id })
+      .then(()=>{
+        this.setCurTrashNote()
+        this.$router.replace({
+            path:'/trash',
+            query:{
+              noteId:this.curTrashNote.id,
+            }
+          })
+      })
     }
   },
